fix warnings

+++ include/board.h
@@ -17,8 +17,8 @@ void board_destroy(Board *b);
 /* 基础属性查询 */
 int board_width(const Board *b);
 int board_height(const Board *b);
-int board_cell_count(const Board *b);   /* width * height */
-int board_filled_count(const Board *b); /* 当前已占用格子数 */
+size_t board_cell_count(const Board *b);   /* width * height */
+size_t board_filled_count(const Board *b); /* 当前已占用格子数 */
 
 /* 边界与占用检查 */
 bool board_in_bounds(const Board *b, int x, int y);
@@ -27,8 +27,8 @@ bool board_is_empty(const Board *b, int x, int y);
 /* 放置与移除：
  * - board_place：若 4 个目标格均在界内且为空，则以给定 mark 填入并返回 true；否则不修改并返回 false。
  * - board_remove：前置条件——必须与先前一次成功放置的 (x,y,type,rotation,mark) 完全一致；否则行为未定义。 */
-bool board_place(Board *b, int x, int y, TetrominoType type, int rotation, char mark);
-void board_remove(Board *b, int x, int y, TetrominoType type, int rotation);
+bool board_place(Board *b, int x, int y, TetrominoType type, size_t rotation, char mark);
+void board_remove(Board *b, int x, int y, TetrominoType type, size_t rotation);
 
 /* 文本可视化：精确输出 height 行、每行 width 个字符；空格子用 '.'，占用格子输出其 mark；每行以 '\n' 结尾。 */
 void board_print(const Board *b, FILE *out);
